Preallocate a buffer for query rows so we don't have to allocate for each one
This speeds up querying quite a bit.  Unfortunately it's using unsafe code to get around iterator lifetime requirements, by storing a pointer to the buffer in the iterator object.  It would be quite easy to incorrect things with the query now.

Apparently, the forthcoming "lending iterator" will make it easier to do this sort of thing.