Start to add saving and loading of data
There's a curious problem where the zstd decoder doesn't read all the data written by the zstd encoder, leaving one byte that we have to read and ignore.  It's not clear if this is caused by zstd itself, or by BufRead not returning the actual position.

It would be good to figure out what's happening.  However for robustness we probably need to know how big blocks are instead of assuming they start exactly where the previous compressed one ended.