Move the QueryIterator functionality into a Scan object, and get it to iterate over multiple blocks
The blocks are kept in a priority queue, and moved to a live set when their first item is reached in the iteration.

I had to remove the values_array hack that was put in to avoid allocating an array of values to return in each iterator call.  The internal machinery of Scan is too complicated now; in particular, there are many iterators returning rows and passing them to each other.  There's some scope for putting it back in a more encapsulated way.