Use Rc for blocks and fetch them from caches properly.
Still need to clean up adding of blocks to a new segment though: they don't need to be Rc while they still belong to the transaction, and the segment saving code doesn't want to care about it when saving them.  At this point, they should probably be a different type from the immutable, saved Block type.  Immutable Blocks are designed to be immutable (easy to load from disk, efficient for querying), whereas the unsaved blocks should be designed to be mutable (easy to add new data to).